미니맵 구현 (테스트 필요함)

+++ src/game/ui/base.py
@@ -1,18 +1,16 @@
-from abc import abstractmethod, ABCMeta
 from typing import NamedTuple
 
-from pygame import Surface
 from pygame.font import Font
 
 
 class UIPosition(NamedTuple):
-    x: int
-    y: int
+    x: float
+    y: float
 
 
 class UISize(NamedTuple):
-    x: int
-    y: int
+    x: float
+    y: float
 
 
 class RGB(NamedTuple):
